# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EduFlow API Documentation",
        "description": "Swagger OpenApi implementation for EduFlow",
        "contact": {
            "email": "support@eduflow.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/teacher/courses/{courseId}/students": {
            "get": {
                "tags": [
                    "Teacher"
                ],
                "summary": "List students enrolled in a course (Teacher only)",
                "operationId": "2498f16efcbd3d94505a31252bfc8b8f",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of students"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teacher/statistics": {
            "get": {
                "tags": [
                    "Teacher"
                ],
                "summary": "Get teacher course and enrollment statistics",
                "operationId": "f646652ea74545e1c426a4f3e3bd749c",
                "responses": {
                    "200": {
                        "description": "Teacher statistics data"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teacher/courses/{courseId}/groups": {
            "get": {
                "tags": [
                    "Teacher"
                ],
                "summary": "List all auto-assigned groups for a course",
                "operationId": "256b12ccb309b9173c284e46ee1ad102",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of groups"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/teacher/courses/{courseId}/groups/{groupId}/students": {
            "get": {
                "tags": [
                    "Teacher"
                ],
                "summary": "List students within a specific group",
                "operationId": "f5e35da9b1f8e7f3ee814b66b9efcea3",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of students in the group"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wishlist": {
            "get": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Retrieve student's wishlist courses",
                "operationId": "3623a3b40a4fb4addae4953160ec84dc",
                "responses": {
                    "200": {
                        "description": "List of wishlisted courses"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/wishlist/courses/{courseId}": {
            "post": {
                "tags": [
                    "Wishlist"
                ],
                "summary": "Toggle a course in the wishlist",
                "operationId": "ac93d58a2bb9eb058f519731ad0b86b3",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wishlist toggled successfully"
                    },
                    "403": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}